+++ kafka-0.9.x/src/main/scala/monix/kafka/Deserializer.scala
@@ -19,42 +19,57 @@ package monix.kafka
 
 import org.apache.kafka.common.serialization._
 import org.apache.kafka.common.serialization.{Deserializer => KafkaDeserializer}
-import language.existentials
 
 /** Wraps a Kafka `Deserializer`, provided for
   * convenience, since it can be implicitly fetched
   * from the context.
+  *
+  * @param className is the full package path to the Kafka `Deserializer`
+  *
+  * @param classType is the `java.lang.Class` for [[className]]
+  *
+  * @param constructor creates an instance of [[classType]].
+  *        This is defaulted with a `Deserializer.Constructor[A]` function that creates a
+  *        new instance using an assumed empty constructor.
+  *        Supplying this parameter allows for manual provision of the `Deserializer`.
   */
 final case class Deserializer[A](
   className: String,
-  classType: Class[_ <: KafkaDeserializer[A]]) {
+  classType: Class[_ <: KafkaDeserializer[A]],
+  constructor: Deserializer.Constructor[A] = (d: Deserializer[A]) => d.classType.newInstance()) {
 
   /** Creates a new instance. */
   def create(): KafkaDeserializer[A] =
-  classType.newInstance()
+    constructor(this)
 }
 
 object Deserializer {
+
+  /** Alias for the function that provides an instance of
+    * the Kafka `Deserializer`.
+    */
+  type Constructor[A] = (Deserializer[A]) => KafkaDeserializer[A]
+
   implicit val forStrings: Deserializer[String] =
-    Deserializer(
+    Deserializer[String](
       className = "org.apache.kafka.common.serialization.StringDeserializer",
       classType = classOf[StringDeserializer]
     )
 
   implicit val forByteArray: Deserializer[Array[Byte]] =
-    Deserializer(
+    Deserializer[Array[Byte]](
       className = "org.apache.kafka.common.serialization.ByteArrayDeserializer",
       classType = classOf[ByteArrayDeserializer]
     )
 
   implicit val forJavaInteger: Deserializer[java.lang.Integer] =
-    Deserializer(
+    Deserializer[java.lang.Integer](
       className = "org.apache.kafka.common.serialization.IntegerDeserializer",
       classType = classOf[IntegerDeserializer]
     )
 
   implicit val forJavaLong: Deserializer[java.lang.Long] =
-    Deserializer(
+    Deserializer[java.lang.Long](
       className = "org.apache.kafka.common.serialization.LongDeserializer",
       classType = classOf[LongDeserializer]
     )