+++ kafka-0.9.x/src/main/scala/monix/kafka/Serializer.scala
@@ -19,42 +19,57 @@ package monix.kafka
 
 import org.apache.kafka.common.serialization._
 import org.apache.kafka.common.serialization.{Serializer => KafkaSerializer}
-import language.existentials
 
 /** Wraps a Kafka `Serializer`, provided for
   * convenience, since it can be implicitly fetched
   * from the context.
+  *
+  * @param className is the full package path to the Kafka `Serializer`
+  *
+  * @param classType is the `java.lang.Class` for [[className]]
+  *
+  * @param constructor creates an instance of [[classType]].
+  *        This is defaulted with a `Serializer.Constructor[A]` function that creates a
+  *        new instance using an assumed empty constructor.
+  *        Supplying this parameter allows for manual provision of the `Serializer`.
   */
 final case class Serializer[A](
   className: String,
-  classType: Class[_ <: KafkaSerializer[A]]) {
+  classType: Class[_ <: KafkaSerializer[A]],
+  constructor: Serializer.Constructor[A] = (s: Serializer[A]) => s.classType.newInstance()) {
 
   /** Creates a new instance. */
   def create(): KafkaSerializer[A] =
-    classType.newInstance()
+    constructor(this)
 }
 
 object Serializer {
+
+  /** Alias for the function that provides an instance of
+    * the Kafka `Serializer`.
+    */
+  type Constructor[A] = (Serializer[A]) => KafkaSerializer[A]
+
   implicit val forStrings: Serializer[String] =
-    Serializer(
+    Serializer[String](
       className = "org.apache.kafka.common.serialization.StringSerializer",
       classType = classOf[StringSerializer]
     )
 
   implicit val forByteArray: Serializer[Array[Byte]] =
-    Serializer(
+    Serializer[Array[Byte]](
       className = "org.apache.kafka.common.serialization.ByteArraySerializer",
       classType = classOf[ByteArraySerializer]
     )
 
   implicit val forJavaInteger: Serializer[java.lang.Integer] =
-    Serializer(
+    Serializer[java.lang.Integer](
       className = "org.apache.kafka.common.serialization.IntegerSerializer",
       classType = classOf[IntegerSerializer]
     )
 
   implicit val forJavaLong: Serializer[java.lang.Long] =
-    Serializer(
+    Serializer[java.lang.Long](
       className = "org.apache.kafka.common.serialization.LongSerializer",
       classType = classOf[LongSerializer]
     )