+++ CHANGES.md
@@ -1,3 +1,28 @@
+## Version 0.14 (May 3, 2017)
+
+- Updates Monix to 2.3.0
+- Updates Kafka 10 to 0.10.2.1
+- Updates Scala to 2.11.11 and 2.12.2
+
+## Version 0.13 (Mar 27, 2017)
+
+- Updates Monix to 2.2.4
+- Activate Scala 2.10 and 2.12 cross-builds
+
+## Version 0.12 (Mar 10, 2017)
+
+- Updates Monix to 2.2.3
+
+## Version 0.11 (Feb 22, 2017)
+
+- Updates Monix to 2.2.2
+
+## Version 0.9 (Jan 25, 2016)
+
+- Updates Monix to 2.2.0
+- [Issue #5](https://github.com/monix/monix-kafka/pull/5):
+  added new optional param `classInstance` to `Serializer` and `Deserializer` 
+
 ## Version 0.8 (Dec 19, 2016)
 
 - Updates Monix to 2.1.2